feat: move serializeAnnotations function to utils and update imports
