+++ packages/bruno-lang/v2/src/jsonToBru.js
@@ -1,26 +1,10 @@
 const _ = require('lodash');
 
-const { indentString, getValueString, getKeyString, getValueUrl } = require('./utils');
+const { indentString, getValueString, getKeyString, getValueUrl, serializeAnnotations } = require('./utils');
 const jsonToExampleBru = require('./example/jsonToBru');
 
 const enabled = (items = [], key = 'enabled') => items.filter((item) => item[key]);
 const disabled = (items = [], key = 'enabled') => items.filter((item) => !item[key]);
-
-const serializeAnnotations = (annotations) => {
-  if (!annotations?.length) return '';
-  return (
-    annotations
-      .map((a) => {
-        if (a.value === undefined) return `@${a.name}`;
-        if (a.value.includes('\n')) {
-          return `@${a.name}('''\n${indentString(a.value)}\n''')`;
-        }
-        const quote = a.value.includes('\'') ? '"' : '\'';
-        return `@${a.name}(${quote}${a.value}${quote})`;
-      })
-      .join('\n') + '\n'
-  );
-};
 
 // remove the last line if two new lines are found
 const stripLastLine = (text) => {
